[mlir-hlo] Remove dead code, fix a few compiler warnings.
PiperOrigin-RevId: 479294037

+++ tensorflow/compiler/xla/mlir_hlo/lib/Dialect/mhlo/transforms/mhlo_canonicalize_scatter.cc
@@ -125,7 +125,7 @@ TypedValue<TensorType> ensureIndexVectorDimPosition(
     indices = insertDegenerateDimensions(b, loc, indices, {indicesRank});
   } else if (indexVectorDim != indicesRank - 1) {
     SmallVector<int64_t> permutation;
-    for (size_t i = 0; i < indicesRank; ++i)
+    for (int64_t i = 0; i < indicesRank; ++i)
       if (i != indexVectorDim) permutation.push_back(i);
     permutation.push_back(indexVectorDim);
     indices = b.create<TransposeOp>(loc, indices,
